pyproject.toml
[tool.poetry]
name = "jira-report-generator"
version = "0.1.0"
description = ""
authors = ["Wonyeong Choi <coma01155@gmail.com>"]
readme = "README.md"
packages = [{include = "jira_report_generator"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
black = "^22.12.0"
pandas = "^1.5.3"
openpyxl = "^3.0.10"
python-dotenv = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
